Something changes

+++ app/Livewire/Dashboard/Profile/ProfileBasic.php
@@ -3,6 +3,8 @@
 namespace App\Livewire\Dashboard\Profile;
 
 use App\Http\Controllers\EditProfile;
+use App\Livewire\Dashboard\Chats\ChatCreate;
+use App\Livewire\Dashboard\Chats\ChatList;
 use App\Livewire\Dashboard\Partials\HeaderUserDropdown;
 use App\Livewire\User;
 use Carbon\Carbon;
@@ -17,26 +19,46 @@ class ProfileBasic extends Component
 {
     use WithFileUploads;
 
-    protected $listeners = [
-        "swalS",
-        "alertSuccess",
-        "alertError",
-        "alertInfo",
-        'fileUpload' => 'update',
-        'refresh-me' => '$refresh',
-    ];
-    public $name, $phone, $address, $email, $username, $avatar, $user;
+    public $name, $description, $email, $username, $avatar, $user;
 
     public function mount()
     {
         $this->name = auth()->user()->name;
         // $this->avatar = auth()->user()->avatar;
-        $this->phone = auth()->user()->phone;
-        $this->address = auth()->user()->address;
+        $this->description = auth()->user()->description;
         $this->email = auth()->user()->email;
         $this->username = auth()->user()->username;
         $this->user = auth()->user();
     }
+
+    public function rules()
+    {
+        return [
+            'name' => ['required', 'min:3', 'max:50'],
+            'avatar' => 'nullable|file|image|mimes:jpg,jpeg,png|max:3069',
+            'username' => [
+                'required', 
+                'min:3', 
+                'max:30', 
+                'regex:/^[a-z][a-z0-9_]*[a-z0-9]+$/',
+                Rule::unique('users')->ignore(auth()->user()->id)
+            ],
+            'description' => ['nullable', 'max:100'],
+        ];
+    }
+
+    protected $messages = [
+        'name.required' => 'Nama harus diisi.',
+        'name.min' => 'Nama yang anda masukan terlalu pendek, minimal 3 karakter.',
+        'name.max' => 'Nama yang anda masukan terlalu panjang, maksimal hanya 50 karakter.',
+        'avatar.mimes' => 'File yang kamu upload harus berformat image (PNG, JPG, JPEG).',
+        'username.required' => 'Username tidak boleh kosong, username harus diisi.',
+        'username.min' => 'Username terlalu pendek, minimal itu hanya 3 karakter.',
+        'username.max' => 'Username terlalu panjang, maksimal itu hanya 20 karakter.',
+        'username.regex' => 'Format username tidak valid, format yang valid: a-z / 0-9 / _ /',
+        'username.unique' => 'Username ini tidak tersedia, silahkan ganti ke yang lain.',
+        'description.max' => 'Deskripsi terlalu panjang, maksimal itu hanya 100 karakter.',
+    ];
 
     public function render()
     {
@@ -47,16 +69,7 @@ class ProfileBasic extends Component
     public function update()
     {
         // // Langkah 2: Validasi data masukan
-        $validatedData = $this->validate([
-            'name' => 'required',
-            'avatar' => 'nullable|file|image|mimes:jpg,jpeg,png,svg|max:2048',
-            'username' => ['required', Rule::unique('users')->ignore(auth()->user()->id)],
-            'phone' => 'nullable',
-            'address' => 'nullable',
-        ], [
-            'avatar.mimes' => 'File yang kamu upload harus berformat image (PNG, JPG, JPEG).',
-            'username.unique' => 'Username ini tidak tersedia, silahkan ganti ke yang lain.'
-        ]);
+        $validatedData = $this->validate($this->rules(), $this->messages);
 
         // dd($validatedData);
         // dd($this->user->avatar);
@@ -68,8 +81,12 @@ class ProfileBasic extends Component
             
             // Jika user memiliki avatar lama, hapus dari storage
             if ($this->user->avatar) {
-                $oldAvatarPath = str_replace('storage/', '', $this->user->avatar); 
-                Storage::delete($oldAvatarPath);
+                // Cek apakah avatar lama berasal dari Google
+                if (strpos($this->user->avatar, 'lh3.googleusercontent.com') === false) {
+                    // Jika bukan dari Google, hapus avatar lama dari storage
+                    $oldAvatarPath = str_replace('storage/', '', $this->user->avatar); 
+                    Storage::delete($oldAvatarPath);
+                }
             }
         } else {
             // Jika tidak ada avatar baru yang diunggah, hapus `avatar` dari `$validatedData`
@@ -79,10 +96,11 @@ class ProfileBasic extends Component
     
         // Langkah 4: Perbarui data pengguna
         $this->user->update($validatedData);
-        $this->dispatch('user-updated');
+
+        $this->dispatch('user-updated')->to(HeaderUserDropdown::class);
         $this->dispatch('refresh')->to(ChatList::class);
         $this->dispatch('refresh')->to(ChatCreate::class);
-        // $this->dispatch('user-updated')->to(HeaderUserDropdown::class);
+        
         $this->dispatch('alert', [
             'title' => 'Berhasil',
             'message' => 'Data Profile berhasil diperbarui',
@@ -94,9 +112,9 @@ class ProfileBasic extends Component
     public function resetField()
     {
         $this->name = auth()->user()->name;
-        $this->phone = auth()->user()->phone;
-        $this->address = auth()->user()->address;
+        $this->description = auth()->user()->description;
         $this->email = auth()->user()->email;
         $this->username = auth()->user()->username;
+        $this->avatar = null;
     }
 }
